bridgeit services client js api
clients can now return a promise from the onSessionExpiry callback and
BridgeIt will wait for the resolution or rejection of that promise
before calling disconnect. This will allow clients to continue
accessing authentication state in asynchronous code in the
onSessionExpiry callback before the auth state is removed in the
disconnect() call.

- example promise onSessionExpiry added to test